refactor: change namespace to be easier
