+++ scripts/Settings/General.cs
@@ -1,7 +1,7 @@
 ﻿using Godot;
 using Newtonsoft.Json.Linq;
 
-namespace RossoSkies1.scripts.Options
+namespace RossoSkies1.scripts.Settings
 {
     internal class General : Options
     {
@@ -27,5 +27,7 @@ namespace RossoSkies1.scripts.Options
                 DisplayServer.WindowSetMode(mode);
             }
         }
+
+
     }
 }
